requirements.txt
asgiref==3.6.0
autopep8==2.0.2
Django==4.1.7
django-cors-headers==3.14.0
djangorestframework==3.14.0
djangorestframework-jwt==1.11.0
djangorestframework-simplejwt==5.2.2
Pillow==9.4.0
pycodestyle==2.10.0
PyJWT==1.7.1
pytz==2022.7.1
sqlparse==0.4.3
tzdata==2022.7
gunicorn==20.1.0
